Fixes and UI improvements.
- New splash screen animation for Android >= API 31.
- Libraries updated.

+++ lib/model/position.dart
@@ -2,7 +2,7 @@ class RecordedPosition {
   double? latitude;
   double? longitude;
   double? altitude;
-  double? timestamp;
+  int? timestamp;
 
   RecordedPosition(this.latitude, this.longitude, this.altitude, this.timestamp);
 }
